requirements.txt
python-docx>=1.1.0
click>=8.1.0
jsonschema>=4.10.0
Pillow>=10.0.0
matplotlib>=3.8.0
geopandas>=0.14.0
shapely>=2.0.0
streamlit>=1.30.0
streamlit-authenticator>=0.4.0
PyMuPDF>=1.24.0
folium>=0.19.0
streamlit-folium>=0.25.0
docx2pdf>=0.1.8
